+++ src/command/hash.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::{
-    database::{ArcBytes, QuickList, Value},
+    database::{arcbytes::ArcBytes, quicklist::QuickList, types::Value},
     engine::engine::StorageEngine,
     error::StoreError,
 };