+++ src/command/zset.rs
@@ -3,7 +3,7 @@ use std::collections::{BTreeMap, HashMap, HashSet};
 use ordered_float::OrderedFloat;
 
 use crate::{
-    database::{ArcBytes, QuickList, Value},
+    database::{arcbytes::ArcBytes, quicklist::QuickList, types::Value},
     engine::engine::StorageEngine,
     error::StoreError,
 };